adding, deleting, sorting and titling of fields work

+++ ToSic.Eav.ManagementUI/EAV/ng/content-type/content-type-field-services.js
@@ -1,12 +1,21 @@
 angular.module('ContentTypeFieldServices', ['ng', 'eavNgSvcs', 'eavGlobalConfigurationProvider'])
-    .factory('contentTypeFieldSvc', function ($http, eavGlobalConfigurationProvider, svcCreator, eavManagementSvc) {
+    .factory('contentTypeFieldSvc', function($http, eavGlobalConfigurationProvider, svcCreator, eavManagementSvc) {
         // start with a basic service which implement the live-list functionality
-        var svc = svcCreator.implementLiveList(function liveListRetrieve() {
-            return $http.get('eav/contenttype/getfields/', { params: { "appid": svc.appId, "staticName": svc.contentType.StaticName } });
-        });
-
-        svc.appId = 0;
+            var svc = {};
+            svc.appId = 0;
             svc.contentType = null;
+            svc.liveListRetrieve = function liveListRetrieve() {
+                return $http.get('eav/contenttype/getfields/', { params: { "appid": svc.appId, "staticName": svc.contentType.StaticName } });
+            }
+            svc.typeListRetrieve = function typeListRetrieve() {
+                return $http.get('eav/contenttype/datatypes/', { params: { "appid": svc.appId } });
+            };
+
+            var svc = angular.extend(svc, svcCreator.implementLiveList(svc.liveListRetrieve));
+            
+            svc.types = svcCreator.implementLiveList(svc.typeListRetrieve);
+
+
 
             svc.moveUp = function moveUp(item) {
                 return $http.get('eav/contenttype/reorder', { params: { appid: svc.appId, contentTypeId: svc.contentType.Id, attributeId: item.Id, direction: 'up' } })
@@ -22,25 +31,45 @@ angular.module('ContentTypeFieldServices', ['ng', 'eavNgSvcs', 'eavGlobalConfigu
                     .then(svc.liveListReload);
             }
 
-            svc.add = function add(item) {
-                return $http.post('eav/contenttype/addfield/', item, { params: { appid: svc.appId, contentTypeId: svc.contentType.Id } })
+            svc.addMany = function add(items, count) {
+                return $http.get('eav/contenttype/addfield/', { params: items[count] })
+                    .then(function() {
+                        if (items.length == ++count)
+                            svc.liveListReload()
+                        else
+                            svc.addMany(items, count);
+                    });
+            };
+
+            svc.add = function addOne(item) {
+                return $http.get('eav/contenttype/addfield/', { params: item })
                     .then(svc.liveListReload);
             };
 
             svc.newItem = function newItem() {
                 return {
+                    AppId: svc.appId,
+                    ContentTypeId: svc.contentType.Id,
                     Id: 0,
                     Type: "String",
                     StaticName: "",
-                    IsTitle: false
+                    IsTitle: svc.liveList().length == 0,
+                    SortOrder: svc.liveList().length
                 }
             };
 
 
             svc.delete = function del(item) {
+                if (item.IsTitle)
+                    throw "Can't delete Title";
                 return $http.delete('eav/contenttype/deletefield', { params: { appid: svc.appId, contentTypeId: svc.contentType.Id, attributeId: item.Id } })
                     .then(svc.liveListReload);
             };
+
+            svc.setTitle = function setTitle(item) {
+                return $http.get('eav/contenttype/setTitle', { params: { appid: svc.appId, contentTypeId: svc.contentType.Id, attributeId: item.Id } })
+                    .then(svc.liveListReload);
+            }
 
 
             return svc;
